refactor: eliminate styled components
