+++ web-client/src/components/figma/BlockStyles.ts
@@ -2,8 +2,31 @@
 
 import styled from 'styled-components';
 
-const ExampleDivStyle = styled.div`
-  margin: 15px;
+const LogoWrapper = styled.div`
+  display: flex;
+  justify-content: space-around;
 `;
 
-export { ExampleDivStyle };
+const StepWrapper = styled('div')`
+  display: flex;
+  flex-direction: column;
+  margin-top: 25px;
+`;
+
+const Explanation = styled('div')`
+  margin-bottom: 50px;
+  margin-top: 30px;
+  display: flex;
+  justify-content: space-around;
+  text-align: center;
+`;
+
+const LoginButtonContainer = styled.div`
+  padding-top: 80px;
+  display: flex;
+  flex-direction: column;
+  width: 220px;
+  margin-bottom: 25px;
+`;
+
+export { Explanation, LoginButtonContainer, LogoWrapper, StepWrapper };
