+++ web-client/src/modules/login/components/LoginSteps/Step3_Authentication.tsx
@@ -1,18 +1,14 @@
-import { Button } from 'antd';
-import React, { useState } from 'react';
+import React from 'react';
 import { useTranslation } from 'react-i18next';
 import logoSmall from 'src/assets/logoSmall.png';
-import { H6Font } from 'src/components/figma/';
-import IntroLogo from 'src/components/IntroLogo/IntroLogo';
-import EventTriggeredModal from 'src/components/Modals/EventTriggeredModal';
-import TitleWithAddon from 'src/components/TitleWithAddon/TitleWithAddon';
-import CONSTANTS from 'src/constants';
-import styled from 'styled-components';
+import {
+  H6Font,
+  LoginButtonContainer,
+  LogoWrapper,
+} from 'src/components/figma/';
 
 import FacebookLoginButton from '../FacebookLoginButton/FacebookLoginButton';
 import GoogleLoginButton from '../GoogleLoginButton/GoogleLoginButton';
-
-const { SUPPORT_EMAIL } = CONSTANTS;
 
 const StepAuthentication: React.FC<StepAuthenticationProps> = ({
   onLoginGoogle,
@@ -20,128 +16,20 @@ const StepAuthentication: React.FC<StepAuthenticationProps> = ({
 }): React.ReactElement<StepAuthenticationProps> => {
   const { t } = useTranslation();
 
-  const [showExplanationModal, setShowExplanationModal] = useState<boolean>(
-    false,
-  );
-  const AuthenticationExplanationModal = () => (
-    <EventTriggeredModal
-      visible
-      finishRequestHandler={() => setShowExplanationModal(false)}
-    >
-      <MainQuestion>
-        <TitleWithAddon level={2} alignAddon="50%">
-          {t('login.steps.3_authentication.popup_explanation.main_question')}
-        </TitleWithAddon>
-      </MainQuestion>
-      <MainAnswer>
-        <H6Font>
-          {t('login.steps.3_authentication.popup_explanation.main_answer')}
-        </H6Font>
-      </MainAnswer>
-      <SecondQuestion>
-        {t('login.steps.3_authentication.popup_explanation.second_question')}
-      </SecondQuestion>
-      <SecondAnswer>
-        <H6Font>
-          {t('login.steps.3_authentication.popup_explanation.second_answer')}{' '}
-          <a href={`mailto:${SUPPORT_EMAIL}`}> {SUPPORT_EMAIL} </a>
-        </H6Font>
-      </SecondAnswer>
-    </EventTriggeredModal>
-  );
-
   return (
     <>
-      <IntroLogo src={logoSmall} alt="logo" />
+      <LogoWrapper>
+        <img src={logoSmall} alt="logo" height="125px" width="125px" />
+      </LogoWrapper>
 
       <LoginButtonContainer>
         <H6Font>{t('login.steps.3_authentication.please')}</H6Font>
-        <ModalTrigger onClick={() => setShowExplanationModal(true)}>
-          <H4>
-            {t('login.steps.3_authentication.popup_explanation.main_question')}
-          </H4>
-        </ModalTrigger>
         <GoogleLoginButton onAuthenticate={onLoginGoogle} />
         <FacebookLoginButton onAuthenticate={onLoginFacebook} />
       </LoginButtonContainer>
-      {showExplanationModal && <AuthenticationExplanationModal />}
     </>
   );
 };
-
-const ModalTrigger = styled(Button)`
-  text-align: center;
-  border: 1px solid #250031;
-  box-sizing: border-box;
-  /* card */
-
-  box-shadow: 0px 4px 12px rgba(0, 0, 0, 0.15);
-  border-radius: 3px;
-
-  /* TODO This is a hack */
-  width: 300px;
-  height: 50px;
-  margin: 33px auto 33px auto;
-`;
-
-const MainQuestion = styled.div`
-  text-align: center;
-  margin-bottom: 25px;
-  margin-top: 10px;
-`;
-const MainAnswer = styled.div`
-  text-align: center;
-  margin-bottom: 60px;
-`;
-
-const SecondQuestion = styled.div`
-  /* H4 */
-  text-align: center;
-  margin-bottom: 25px;
-
-  font-family: Roboto;
-  font-style: normal;
-  font-weight: 500;
-  font-size: 18px;
-  line-height: 26px;
-  /* or 144% */
-
-  text-align: center;
-
-  /* @heading-color */
-
-  color: rgba(0, 0, 0, 0.85);
-`;
-
-const SecondAnswer = styled.div`
-  text-align: center;
-  margin-bottom: 120px;
-`;
-
-const H4 = styled.span`
-  /* H4 */
-
-  font-family: Roboto;
-  font-style: normal;
-  font-weight: bold;
-  font-size: 14px;
-  line-height: 22px;
-  /* or 157% */
-
-  text-align: center;
-
-  /* @heading-color */
-
-  color: rgba(0, 0, 0, 0.85);
-`;
-
-const LoginButtonContainer = styled.div`
-  padding-top: 80px;
-  display: flex;
-  flex-direction: column;
-  width: 220px;
-  margin-bottom: 25px;
-`;
 
 interface StepAuthenticationProps {
   onLoginGoogle: Function;
